Move `create_generic_constant` to the `computation_constructing_utils` module.
PiperOrigin-RevId: 259427365
